Add files via upload
First iteration of the DPS calculator. At the moment it is only for melee, and lacking in error handling. 
Some of the math is a bit off when compared to other calculators but will have to fix this later as I might need some help with the debugging

Form1.cs contains the bulk of the code